+++ iOSWebBox/CoconutSDK/Sources/CoconutSDK/Core/ComponentContext.swift
@@ -6,7 +6,7 @@ import WebKit
 public class ComponentContext {
 
     public let applicationContext: UIApplication
-    public var sdkVersion: String = "3.5.0"
+    public var sdkVersion: String = "3.5.1"
 
     /// Shared EventEmitter for native → H5 push.
     public var eventEmitter: EventEmitter!